새로운 background, newAPI, error핸들링, locating feature 마니마니

+++ src/pages/MyPage.tsx
@@ -57,6 +57,7 @@ const MyPage: React.FC = () => {
   };
 
   const deleteSnowman = async (id: number) => {
+  confirm("☃️ : 절 정말 녹이실건가요 주인님 ? 🥲");
     try {
       await deleteMySnowman(id);
       setSnowmans(snowmans.filter(snowman => snowman.id !== id)); // Remove the deleted snowman from the UI
